test(desktop): split Pocket token boundary tests
